+++ backend/src/mongodriver/connectionAndTypes.go
@@ -1,0 +1,25 @@
+package mongodriver
+
+import (
+	"context"
+
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+// Connect connects to the MongoDB and returns the client
+func Connect() (*mongo.Client, error) {
+	// Set client options
+	clientOptions := options.Client().ApplyURI("mongodb://mongo:27017")
+	// Connect to MongoDB
+	client, err := mongo.Connect(context.TODO(), clientOptions)
+	if err != nil {
+		return nil, err
+	}
+	// Check the connection
+	err = client.Ping(context.TODO(), nil)
+	if err != nil {
+		return nil, err
+	}
+	return client, nil
+}
